Neural Net/requirements.txt
numpy
Pillow
sklearn
matplotlib